Barangay data moved to constant folder
